migrate client from CRA to Vite
- Replace react-scripts with Vite for ~99% faster dev server startup

- Remove unused server-side packages (discord.js, cheerio, chalk, ejs, handlebars, node-fetch, chai)

- Rename .js files to .jsx for proper JSX handling

- Add Vitest for testing

- Fix ColorModeContext usage bug in AppBar

Fixes #96
