+++ src/client/package.json
@@ -2,15 +2,16 @@
   "name": "twitchbot-client",
   "version": "1.0.0",
   "description": "the twitchbot client",
-  "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "cross-env ESLINT_NO_DEV_ERRORS=true; react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest",
     "lint:quick": "quick-lint-js ./src",
-    "lint": "eslint src/**/*.js",
-    "lint:fix": "eslint src/**/*.js --fix",
+    "lint": "eslint src/**/*.{js,jsx}",
+    "lint:fix": "eslint src/**/*.{js,jsx} --fix",
     "prettier": "prettier --check .",
     "prettier:fix": "prettier . --write",
     "preflight": "npm run lint && npm run prettier",
@@ -30,51 +31,29 @@
     "@mui/material": "^7.0.2",
     "@tanstack/react-query": "^5.72.2",
     "axios": "^1.8.4",
-    "chai": "^5.2.0",
-    "chalk": "^5.4.1",
-    "cheerio": "^1.0.0",
-    "discord.js": "^14.18.0",
-    "ejs": "^3.1.10",
-    "handlebars": "^4.7.8",
-    "node-fetch": "^3.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router": "^7.5.0",
     "react-router-dom": "^7.5.0",
-    "react-scripts": "^5.0.1",
     "socket.io-client": "^4.8.1",
     "web-vitals": "^4.2.4"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@babel/plugin-transform-private-property-in-object": "^7.25.9",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
-    "cross-env": "^7.0.3",
+    "@vitejs/plugin-react": "^4.3.4",
     "eslint": "^9.24.0",
     "eslint-config-prettier": "^10.1.2",
     "eslint-plugin-jest-dom": "^5.5.0",
     "eslint-plugin-prettier": "^5.2.6",
     "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "jsdom": "^26.0.0",
     "prettier": "^3.5.3",
-    "quick-lint-js": "^3.2.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "jest": {
-    "moduleNameMapper": {
-      "^axios$": "axios/dist/node/axios.cjs"
-    }
+    "quick-lint-js": "^3.2.0",
+    "vite": "^6.0.3",
+    "vitest": "^2.1.8"
   }
 }
