+++ src/client/src/components/ProtectedRoute.jsx
