+++ src/client/src/components/TemplateComponent.jsx
