+++ src/client/src/pages/HelpPage.jsx
